Design update
Working on the overall look and feel, the about section is pretty much
done with the content and the way it looks and feels. I will be working
on the Portfolio section next, This will use a Angular dynamic display
to pull and display all of the data in an appealing and user friendly
way. Will have to work on a lot of the backend code and ideas